Refactor

+++ chess/src/chess/logic/DrawChecker.h
@@ -12,7 +12,7 @@ namespace Chess
 	{
 	private:
 		int m_movesCountWithoutPawnAndTaking = 0;
-		int m_lastCountEatenPeaces = 0;
+		size_t m_lastCountEatenPeaces = 0;
 
 	private:
 		void CalculateMovesCountWithoutPawnAndTaking(const std::shared_ptr<Chessboard>& chessboard);
